+++ ic-agent/src/lib.rs
@@ -103,11 +103,12 @@
 //! generally available, additional details about the versions supported will
 //! be available here.
 
-#![deny(
+#![warn(
     missing_docs,
     rustdoc::broken_intra_doc_links,
     rustdoc::private_intra_doc_links
 )]
+#![cfg_attr(not(target_family = "wasm"), warn(clippy::future_not_send))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 #[cfg(all(feature = "hyper", target_family = "wasm"))]
@@ -117,11 +118,9 @@ pub mod agent;
 pub mod export;
 pub mod identity;
 
+use agent::response_authentication::LookupPath;
 #[doc(inline)]
-pub use agent::{
-    agent_error, agent_error::AgentError, response_authentication::lookup_value, Agent,
-    NonceFactory, NonceGenerator,
-};
+pub use agent::{agent_error, agent_error::AgentError, Agent, NonceFactory, NonceGenerator};
 #[doc(inline)]
 pub use ic_transport_types::{to_request_id, RequestId, RequestIdError};
 #[doc(inline)]
@@ -129,3 +128,13 @@ pub use identity::{Identity, Signature};
 
 // Re-export from ic_certification for backward compatibility.
 pub use ic_certification::{hash_tree, Certificate};
+
+/// Looks up a value in the certificate's tree at the specified hash.
+///
+/// Returns the value if it was found; otherwise, errors with `LookupPathAbsent`, `LookupPathUnknown`, or `LookupPathError`.
+pub fn lookup_value<P: LookupPath, Storage: AsRef<[u8]>>(
+    tree: &ic_certification::certificate::Certificate<Storage>,
+    path: P,
+) -> Result<&[u8], AgentError> {
+    agent::response_authentication::lookup_value(&tree.tree, path)
+}