+++ ic-utils/src/lib.rs
@@ -1,12 +1,13 @@
 //! ic-utils is a collection of utilities to help build clients and canisters running
 //! on the Internet Computer. It is meant as a higher level tool.
 
-#![deny(
+#![warn(
     missing_docs,
     missing_debug_implementations,
     rustdoc::broken_intra_doc_links,
     rustdoc::private_intra_doc_links
 )]
+#![cfg_attr(not(target_family = "wasm"), warn(clippy::future_not_send))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 /// Utilities to encapsulate calls to a canister.