+++ ref-tests/tests/integration.rs
@@ -100,17 +100,18 @@ fn canister_reject_call() {
 
         let result = alice.wallet_send(*bob.canister_id(), 1_000_000).await;
 
-        assert_eq!(
+        assert!(matches!(
             result,
             Err(AgentError::ReplicaError(RejectResponse {
                 reject_code: RejectCode::DestinationInvalid,
-                reject_message: format!(
-                    "Canister {} has no update method 'wallet_send'",
-                    alice.canister_id()
-                ),
-                error_code: None
-            }))
-        );
+                reject_message,
+                error_code: None,
+                ..
+            })) if reject_message == format!(
+                "Canister {} has no update method 'wallet_send'",
+                alice.canister_id()
+            )
+        ));
 
         Ok(())
     });
@@ -547,7 +548,7 @@ mod sign_send {
                 .await?;
 
             assert!(
-                matches!(response, RequestStatusResponse::Replied(ReplyResponse { arg: result }) if result == b"hello")
+                matches!(response, RequestStatusResponse::Replied(ReplyResponse { arg: result, .. }) if result == b"hello")
             );
             Ok(())
         })